.emacs: auto-load ELPA packages
It checks if ~/.emacs.d/elpa/ exists and is a directory. If not, it
calls package-refresh-contents to download the index of packages
from ELPA and MELPA. Then call package-install to ensure that all my
selected packages are installed. If any packages are not installed,
the necessary ones will be downloaded and installed.