.bashrc: pretty up welcome MOTD; add some checks
The welcome message now actually has "Welcome" in it, provided by
figlet's large text output. This and a fortune cookie is piped
together to cowsay which is in turn piped through lolcat to make it
rainbow-y.

Checks were also added to see if cowsay, figlet, and fortune were
installed. If they are installed, proceed as normal; if any aren't,
notify the user. An additional check was added to check for lolcat.
If lolcat isn't found, we can just output without piping to it.