.gitconfig: add flags to make Emacs run in terminal
On my GUI systems, I prefer to have both the GUI and terminal
versions of Emacs installed. Because of this, Emacs will open up as
the out-of-terminal GUI version when started from the terminal. To
remedy this, the flags to run in no-window mode are added to the
.gitconfig.

To my knowledge, the redundancy on non-GUI systems (or SSH
connections) seems to have no effects.