custom.el: use konsole foreground/background font colors
I tty mode, emacs just inherits the foreground and background colors for the
default font face from the console. But in GUI mode, it uses the default emacs
colors (light mode). This way, emacs looks the same whether it's in tty or GUI
mode (at least, in terms of faces).