offload backing up of tarballs to separate script
Putting the original copy_to_waiter function into the background caused some
display oddities (and maybe performance slowdowns). Additionally, if both
auto-ripper scripts are uploading at the same time, it kills the upload
speed. This means it takes longer to start the next rip. It's easier to just
serialize the upload and have it be handled by a different script entirely.